Modifying TempDao and TempRepo functions into suspend functions. Adding Temperature chart in FertilityFragment.

+++ app/src/main/java/com/benlefevre/monendo/data/models/Temperature.kt
@@ -7,7 +7,7 @@ import java.util.*
 
 @Entity
 data class Temperature(
-    @ColumnInfo(name = "value") var value : Double,
+    @ColumnInfo(name = "value") var value : Float,
     @ColumnInfo(name = "date") var date: Date
 ) {
     @PrimaryKey(autoGenerate = true)
